+++ src/worker.js
@@ -1,5 +1,7 @@
 'use strict';
 
+process.env.CLASSIC_TRIP_PROCESS_ROLE = process.env.CLASSIC_TRIP_PROCESS_ROLE || 'worker';
+
 const { env } = require('./config/env');
 const { connectDb, mongoose } = require('./config/db');
 const { connectRedis, closeRedis } = require('./config/redis');